#8610 Add INachrichtenService and move NachrichtenDTO to DTOs
